Refactor create release form state
