+++ web/features/deployments/create-release/state/index.ts
@@ -1,14 +1,78 @@
 'use client'
 
+import type { SourceAppPickerValue } from '../ui/source-app-picker-value'
 import type { UnsupportedDslNode } from '../../shared/domain/error'
-import type { CreateReleaseFormValues, ReleaseSourceMode } from './types'
 import { atom, useAtomValue } from 'jotai'
-import { encodeDslContent, isWorkflowDsl } from '../../shared/domain/dsl'
 import {
-  createReleaseDslFileFieldAtom,
-  createReleaseSourceAppFieldAtom,
-  createReleaseSourceModeFieldAtom,
-} from './use-create-release-form'
+  atomWithForm,
+  createFormAtoms,
+} from 'jotai-tanstack-form'
+import * as z from 'zod'
+import { encodeDslContent, isWorkflowDsl } from '../../shared/domain/dsl'
+
+export type ReleaseSourceMode = 'sourceApp' | 'dsl'
+
+export type CreateReleaseFormValues = {
+  releaseSourceMode: ReleaseSourceMode
+  sourceApp?: SourceAppPickerValue
+  dslFile?: File
+  releaseName: string
+  releaseDescription: string
+}
+
+export const DEFAULT_CREATE_RELEASE_FORM_VALUES: CreateReleaseFormValues = {
+  releaseSourceMode: 'sourceApp',
+  sourceApp: undefined,
+  dslFile: undefined,
+  releaseName: '',
+  releaseDescription: '',
+}
+
+export const RELEASE_NAME_REQUIRED_ERROR = 'releaseNameRequired'
+
+const createReleaseFormSchema = z.object({
+  releaseSourceMode: z.union([z.literal('sourceApp'), z.literal('dsl')]),
+  sourceApp: z.custom<CreateReleaseFormValues['sourceApp']>().optional(),
+  dslFile: z.custom<CreateReleaseFormValues['dslFile']>().optional(),
+  releaseName: z.string().trim().min(1, RELEASE_NAME_REQUIRED_ERROR),
+  releaseDescription: z.string(),
+})
+
+type CreateReleaseSubmit = (value: CreateReleaseFormValues) => Promise<void> | void
+
+type CreateReleaseSubmitMeta = {
+  createRelease: CreateReleaseSubmit
+}
+
+const noopCreateRelease: CreateReleaseSubmit = () => undefined
+
+export const createReleaseFormAtom = atomWithForm({
+  defaultValues: DEFAULT_CREATE_RELEASE_FORM_VALUES,
+  onSubmitMeta: {
+    createRelease: noopCreateRelease,
+  },
+  validators: {
+    onSubmit: createReleaseFormSchema,
+  },
+  onSubmit: ({ value, meta }) => meta.createRelease(value),
+})
+
+const createReleaseFormAtoms = createFormAtoms(createReleaseFormAtom)
+
+export const createReleaseFormStateAtom = createReleaseFormAtoms.stateAtom
+export const createReleaseFormValuesAtom = createReleaseFormAtoms.valuesAtom
+export const createReleaseFormIsSubmittingAtom = createReleaseFormAtoms.isSubmittingAtom
+export const setCreateReleaseFormFieldAtom = createReleaseFormAtoms.setFieldAtom
+export const createReleaseSourceModeFieldAtom = createReleaseFormAtoms.fieldAtom('releaseSourceMode')
+export const createReleaseSourceAppFieldAtom = createReleaseFormAtoms.fieldAtom('sourceApp')
+export const createReleaseDslFileFieldAtom = createReleaseFormAtoms.fieldAtom('dslFile')
+export const createReleaseNameFieldAtom = createReleaseFormAtoms.fieldAtom('releaseName')
+export const createReleaseDescriptionFieldAtom = createReleaseFormAtoms.fieldAtom('releaseDescription')
+export const submitCreateReleaseFormAtom = atom(null, (get, _set, createRelease: CreateReleaseSubmit) => {
+  const form = get(createReleaseFormAtom)
+
+  return form.api.handleSubmit({ createRelease } satisfies CreateReleaseSubmitMeta)
+})
 
 type CreateReleaseConfig = {
   appInstanceId: string
